backend/src/requirements.txt
asgiref==3.6.0
Django==4.1.7
django-cors-headers==3.14.0
django-extensions==3.2.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
numpy==1.24.2
opencv-python==4.7.0.72
pytz==2022.7.1
scipy==1.10.1
sqlparse==0.4.3
tzdata==2022.7
